# Process `MemPostings.Delete()` with `GOMAXPROCS` workers