We are still seeing lock contention on `MemPostings.mtx` for customers with lots of labels per series, and `MemPostings.Delete()` is by far the most expensive operation on that mutex.

This adds parallelism to that method, trying to reduce the amount of time we spend with the mutex held.

I updated the benchmark to create more series to get a more meaningful result.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb/index
cpu: Apple M3 Pro
                                            │   main_2m    │             parallel_2m             │
                                            │    sec/op    │    sec/op     vs base               │
MemPostings_Delete/refs=1/readers=0-12        46.51m ± 10%   24.65m ±  4%  -47.00% (p=0.002 n=6)
MemPostings_Delete/refs=1/readers=1-12        46.23m ±  2%   24.00m ±  8%  -48.08% (p=0.002 n=6)
MemPostings_Delete/refs=1/readers=10-12       46.58m ±  7%   24.67m ± 11%  -47.04% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=0-12      68.75m ±  7%   36.97m ±  4%  -46.23% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=1-12      71.69m ±  9%   36.25m ±  5%  -49.43% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=10-12     71.19m ±  6%   36.64m ±  3%  -48.53% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=0-12    68.39m ± 11%   56.41m ±  8%  -17.52% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=1-12    69.07m ± 13%   58.22m ±  9%  -15.71% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=10-12   70.15m ±  4%   58.45m ±  9%  -16.68% (p=0.002 n=6)
geomean                                       60.97m         37.24m        -38.93%
```

I recommend checking the diff hiding the whitespace changes.