@colega 2024-09-26
@bboreham I've extracted a function, PTAL. Benchmark results didn't change.